+++ Accessdb.h
@@ -15,7 +15,7 @@ class AccessRec {
  public:
 	key128_t  m_key128;
 	long      m_ip;
-	long long m_fbId; // facebook id, 0 if none
+	int64_t m_fbId; // facebook id, 0 if none
 };
 
 class Accessdb {
@@ -31,8 +31,8 @@ class Accessdb {
 
 	Rdb *getRdb() { return &m_rdb; }
 
-	key128_t makeKey1 ( long long now, long long widgetId64 ) ;
-	key128_t makeKey2 ( long long now, long long widgetId64 ) ;
+	key128_t makeKey1 ( int64_t now, int64_t widgetId64 ) ;
+	key128_t makeKey2 ( int64_t now, int64_t widgetId64 ) ;
 
 	bool addAccess ( class HttpRequest *r , long ip );
